requirements.txt
asgiref==3.2.10
astroid==2.4.1
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==3.1.2
django-crispy-forms==1.9.2
django-heroku==0.3.1
dnspython==1.16.0
Flask==1.1.1
Flask-PyMongo==2.3.0
Flask-WTF==0.14.2
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.0.0
pipenv==2018.11.26
psycopg2==2.7.5
psycopg2-binary==2.8.6
pylint==2.5.2
pymongo==3.9.0
python-dotenv==0.14.0
pytz==2018.5
requests==2.22.0
selenium==3.141.0
simplejson==3.17.2
six==1.14.0
soupsieve==1.9.4
sqlparse==0.3.1
stripe==2.55.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.6
virtualenv==16.0.0
virtualenv-clone==0.5.4
Werkzeug==0.16.0
whitenoise==5.2.0
wrapt==1.12.1
WTForms==2.2.1
